chore: create confetti burst
