+++ modules/i18n/kreyol/hk.json
@@ -726,6 +726,7 @@
     "findTitle": "Chèche Tout Dosye",
     "findDescription": "Chèche nenpot rezidan oswa byen. Sipèvizè itilize li pou idèntifye tandans ak pran desizyon baze sou prèv.",
     "settingsTitle": "Prefèrans Ou",
-    "settingsDescription": "Chanje lang, aktive mòd nwa, epi jère kont ou — tout nan yon sèl kote."
+    "settingsDescription": "Chanje lang, aktive mòd nwa, epi jère kont ou — tout nan yon sèl kote.",
+    "dismiss": "Femen konsèy"
   }
 }
